requirements.txt
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.8
PyYAML==6.0.1 